# docs/swagger.yaml
basePath: /
definitions:
...
  entities.Booking:
    properties:
      app_code:
        type: string
      appointment_id:
        type: string
      attendee_count:
        type: integer
      available:
        type: boolean
      booking_code:
        description: Permanent booking code for all bookings
        type: string
      capacity:
        type: integer
      created_at:
        type: string
      date:
        type: string
      deleted_at:
        format: date-time
        type: string
      description:
        description: Additional info from the booker
        type: string
      email:
        type: string
      end_time:
        type: string
      id:
        type: string
      is_slot:
        type: boolean
      name:
        type: string
      notification_channel:
        type: string
      notification_status:
        type: string
      phone:
        type: string
      seats_booked:
        type: integer
      start_time:
        type: string
      status:
        description: 'Booking status: active, cancelled, etc.'
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
...
  errors.APIErrorResponse:
    properties:
      code:
        type: string
      fields:
        items:
          $ref: '#/definitions/errors.FieldError'
        type: array
      message:
        type: string
      meta: {}
      request_id:
        type: string
      status:
        type: integer
    type: object
  errors.FieldError:
    properties:
      field:
        type: string
      message:
        type: string
      rule:
        type: string
    type: object
...
  requests.BookingRequest:
    properties:
      app_code:
        type: string
      attendee_count:
        minimum: 1
        type: integer
      date:
        type: string
      description:
        type: string
      device_token:
        type: string
      email:
        type: string
      end_time:
        type: string
      name:
        type: string
      phone:
        type: string
      start_time:
        type: string
    required:
    - app_code
    - date
    - end_time
    - start_time
    type: object
...
  responses.PaginatedResponse:
    properties:
      items:
        items: {}
        type: array
      page:
        type: integer
      per_page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
...
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a comprehensive API for creating and managing appointments.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Asiko API
  version: "1.0"
paths:
...
  /appointments/book:
    post:
      consumes:
      - application/json
      description: Creates a booking for an appointment as a guest user. Name and
        email/phone are required.
      operationId: bookGuestAppointment
      parameters:
      - description: Booking Details
        in: body
        name: booking
        required: true
        schema:
          $ref: '#/definitions/requests.BookingRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/entities.Booking'
        "400":
          description: Invalid request payload or validation error
          schema:
            $ref: '#/definitions/errors.APIErrorResponse'
        "409":
          description: Slot unavailable or capacity exceeded
          schema:
            $ref: '#/definitions/errors.APIErrorResponse'
      summary: Book an appointment (Guest)
      tags:
      - Bookings
  /appointments/book/registered:
    post:
      consumes:
      - application/json
      description: Creates a booking for an appointment as a registered user.
      operationId: bookRegisteredUserAppointment
      parameters:
      - description: Booking Details
        in: body
        name: booking
        required: true
        schema:
          $ref: '#/definitions/requests.BookingRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/entities.Booking'
        "400":
          description: Invalid request payload or validation error
          schema:
            $ref: '#/definitions/errors.APIErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.APIErrorResponse'
        "409":
          description: Slot unavailable or capacity exceeded
          schema:
            $ref: '#/definitions/errors.APIErrorResponse'
      security:
      - BearerAuth: []
      summary: Book an appointment (Registered User)
      tags:
      - Bookings
...
  /appointments/registered:
    get:
      description: Retrieves a paginated list of all bookings made by the currently
        authenticated user.
      operationId: getUserRegisteredBookings
      parameters:
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Page size (default: 10)'
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/responses.PaginatedResponse'
            - properties:
                items:
                  items:
                    $ref: '#/definitions/entities.Booking'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.APIErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/errors.APIErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user's registered bookings
      tags:
      - Bookings
  /appointments/slots/{app_code}:
    get:
      description: Retrieves a paginated list of all available booking slots for a
        given appointment.
      operationId: getAvailableSlots
      parameters:
      - description: Appointment identifier (app_code)
        in: path
        name: app_code
        required: true
        type: string
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Page size (default: 500)'
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/responses.PaginatedResponse'
            - properties:
                items:
                  items:
                    $ref: '#/definitions/entities.Booking'
                  type: array
              type: object
        "400":
          description: Missing appointment code parameter
          schema:
            $ref: '#/definitions/errors.APIErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/errors.APIErrorResponse'
      summary: Get available slots for an appointment
      tags:
      - Bookings
  /appointments/slots/{app_code}/by-day:
    get:
      description: Retrieves a paginated list of available slots for an appointment
        on a specific day.
      operationId: getAvailableSlotsByDay
      parameters:
      - description: Appointment identifier (app_code)
        in: path
        name: app_code
        required: true
        type: string
      - description: Date in YYYY-MM-DD format
        in: query
        name: date
        required: true
        type: string
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Page size (default: 200)'
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/responses.PaginatedResponse'
            - properties:
                items:
                  items:
                    $ref: '#/definitions/entities.Booking'
                  type: array
              type: object
        "400":
          description: Missing or invalid parameters
          schema:
            $ref: '#/definitions/errors.APIErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/errors.APIErrorResponse'
      summary: Get available slots for a specific day
      tags:
      - Bookings
  /appointments/users/{app_code}:
    get:
      description: Retrieves a paginated list of all users/bookings for a specific
        appointment.
      operationId: getUsersRegisteredForAppointment
      parameters:
      - description: Appointment identifier (app_code)
        in: path
        name: app_code
        required: true
        type: string
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Page size (default: 10)'
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/responses.PaginatedResponse'
            - properties:
                items:
                  items:
                    $ref: '#/definitions/entities.Booking'
                  type: array
              type: object
        "400":
          description: Missing appointment code parameter
          schema:
            $ref: '#/definitions/errors.APIErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.APIErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/errors.APIErrorResponse'
      security:
      - BearerAuth: []
      summary: Get all bookings for an appointment
      tags:
      - Appointments
...
  /bookings/{booking_code}:
    delete:
      description: Cancels a booking by its unique booking_code. This is a soft delete.
      operationId: cancelBookingByCode
      parameters:
      - description: Unique Booking Code
        in: path
        name: booking_code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.Booking'
        "400":
          description: Error while cancelling booking
          schema:
            $ref: '#/definitions/errors.APIErrorResponse'
        "404":
          description: Booking not found
          schema:
            $ref: '#/definitions/errors.APIErrorResponse'
      summary: Cancel a booking
      tags:
      - Bookings
    get:
      description: Retrieves booking details by its unique booking_code.
      operationId: getBookingByCode
      parameters:
      - description: Unique Booking Code
        in: path
        name: booking_code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.Booking'
        "400":
          description: Missing booking_code parameter
          schema:
            $ref: '#/definitions/errors.APIErrorResponse'
        "404":
          description: Booking not found
          schema:
            $ref: '#/definitions/errors.APIErrorResponse'
      summary: Get booking by code
      tags:
      - Bookings
    put:
      consumes:
      - application/json
      description: Updates a booking by its unique booking_code. Can be used to reschedule.
      operationId: updateBookingByCode
      parameters:
      - description: Unique Booking Code
        in: path
        name: booking_code
        required: true
        type: string
      - description: New Booking Details
        in: body
        name: booking
        required: true
        schema:
          $ref: '#/definitions/requests.BookingRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.Booking'
        "400":
          description: Invalid request, validation error, or slot not available
          schema:
            $ref: '#/definitions/errors.APIErrorResponse'
        "404":
          description: Booking not found
          schema:
            $ref: '#/definitions/errors.APIErrorResponse'
        "409":
          description: Requested slot not available or capacity exceeded
          schema:
            $ref: '#/definitions/errors.APIErrorResponse'
      summary: Update/Reschedule a booking
      tags:
      - Bookings
  /bookings/{booking_code}/reject:
    post:
      description: Rejects a booking by its unique booking_code. This is a soft delete.
      operationId: rejectBookingByCode
      parameters:
      - description: Unique Booking Code
        in: path
        name: booking_code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.Booking'
        "400":
          description: Error while rejecting booking
          schema:
            $ref: '#/definitions/errors.APIErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.APIErrorResponse'
        "404":
          description: Booking not found
          schema:
            $ref: '#/definitions/errors.APIErrorResponse'
      security:
      - BearerAuth: []
      summary: Reject a booking
      tags:
      - Bookings
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    description: Type 'Bearer' followed by a space and a JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"